dm: address FK hot update comments
